+++ package.json
@@ -1,8 +1,8 @@
 {
-    "name": "node-red-contrib-deebot",
-    "version": "1.0.0",
-    "description": "Node for Ecovacs Deebot vacuum cleaner",
-    "main": "deebot.js",
+    "name": "node-red-contrib-ecovacs-deebot",
+    "version": "0.0.1",
+    "description": "NodeRed node for Ecovacs Deebot vacuum cleaner",
+    "main": "ecovacs-deebot.js",
     "scripts": {
         "test": "echo \"Error: no test specified\" && exit 1"
     },
@@ -10,7 +10,7 @@
     "license": "ISC",
     "node-red": {
         "nodes": {
-            "deebot": "deebot.js"
+            "deebot": "ecovacs-deebot.js"
         }
     },
     "dependencies": {
